requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
asyncpg==0.29.0
attrs==24.1.0
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
frozenlist==1.4.1
idna==3.7
kombu==5.4.0
magic-filter==1.0.12
multidict==6.0.5
prometheus-client==0.16.0
prometheus-summary==0.1.2
prompt_toolkit==3.0.47
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
quantile-estimator==0.1.2
redis==5.0.8
six==1.16.0
SQLAlchemy==2.0.32
transitions==0.9.1
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
langchain_openai
langchain_community==0.2.6
sentence_transformers
langchain_chroma
yandexcloud
requests
